READ.ME
Takes a user inputted date and stock ticker and uses machine learning to make a price prediction of the specified day based on the prior 30 days of closing stock prices.
This was a project based on my personal interests in machine learning and finance. This is a self-taught project, and I am still working on adding features such as a GUI 
and graphing of data.